backend/requirements.txt
# Core Flask Framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Werkzeug==2.3.7

# Database
pymongo==4.5.0
python-dotenv==1.0.0

# Security & Authentication
bcrypt==4.0.1
PyJWT==2.8.0

# Data Processing & Analysis
pandas==2.1.1
numpy==1.26.0
python-dateutil==2.8.2

# Machine Learning & AI
scikit-learn==1.3.0
prophet==1.1.4
scipy==1.11.1

# Stock Market Data
yfinance==0.2.18
pandas-datareader==0.10.0
requests==2.31.0

# Data Visualization
matplotlib==3.8.0
seaborn==0.12.2
plotly==5.17.0

# Image Processing & OCR
opencv-python==4.8.1.78
pytesseract==0.3.10
Pillow==10.0.1

# Additional Utilities
lxml==4.9.3
beautifulsoup4==4.12.2
html5lib==1.1
urllib3==2.0.4
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
